ml_project/requirements.txt
antlr4-python3-runtime==4.8
astroid==2.5
attrs==20.3.0
click==7.1.2
coverage==5.5
hydra-core==1.0.6
importlib-metadata==4.0.0
importlib-resources==5.1.2
iniconfig==1.1.1
isort==5.8.0
joblib==1.0.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
more-itertools==8.7.0
numpy==1.20.2
omegaconf==2.0.6
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
pytest-cov==2.11.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.1
